+++ CHANGES.rst
@@ -233,6 +233,10 @@ astropy.config
 astropy.constants
 ^^^^^^^^^^^^^^^^^
 
+- Add CODATA 2017 constants but not make them default because the
+  redefinition of SI units that will follow has not been implemented
+  yet. [#8167]
+
 astropy.convolution
 ^^^^^^^^^^^^^^^^^^^